Initial set of changes to ptpd-2.4-natinst: updated to run configure separately for each platform, added initial ptp API and libptp, added options for running in a threaded environment and specifying an alternate servo.

+++ src/libcck/cck.h
@@ -4,7 +4,7 @@
  * Copyright (c) 2014 Wojciech Owczarek,
  *
  * All Rights Reserved
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are
  * met:
@@ -13,7 +13,7 @@
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE AUTHORS ``AS IS'' AND ANY EXPRESS OR
  * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -29,7 +29,7 @@
 
 /**
  * @file   cck.h
- * 
+ *
  * @brief  libCCK main header file. Include this to start using libCCK
  *
  */
@@ -56,7 +56,7 @@
 /* Own xmalloc/calloc - printf until we have CckLogTarget component written and can log this proper */
 #define CCKCALLOC(ptr, size) \
         if(!((ptr)=malloc(size))) { \
-                printf("Critical: failed to allocate %ld bytes of memory.memory", size); \
+           printf("Critical: failed to allocate %ld bytes of memory.memory", (long int) size); \
                 exit(1); \
         } \
         memset(ptr, 0, size);
